Moved to OutputAllocator classes to new file
